// src/app/model/assets.ts
import { Result } from './result';

export class Assets extends Result {
    accountBalance: number;
    earning: number;
    earningMoney: number;
    earningRetain: number;
    rewardMoney: number;
    moneyWithdraw: number;
    overdue: number;
    xcbTotalMoney: number;
    eswAccountBalance: number;
    eswEarning: number;
    eswEarningMoney: number;
    eswEarningRetain: number;
    eswOverdue: number;
    eswAvialBalance: number;
}